requirements.txt
APScheduler==3.6.3
bidict==0.21.2
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cryptography==3.4.7
ecdsa==0.14.1
fastapi==0.65.1
fastapi-jwt-auth==0.5.0
gunicorn==20.0.4
h11==0.10.0
h2==4.0.0
hpack==4.0.0
httptools==0.2.0
hyperframe==6.0.0
idna==2.10
priority==1.3.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
PyJWT==1.7.1
python-dotenv==0.14.0
python-engineio==4.2.0
pytz==2020.1
requests==2.24.0
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.3.19
starlette==0.14.2
toml==0.10.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.10
uvicorn==0.13.4
uvloop==0.15.2
wsproto==0.15.0
